requirements.txt
streamlit==1.28.0
pillow==10.0.0
numpy==1.24.0
google-generativeai==0.3.1
transformers==4.34.0
torch==2.0.1
sentencepiece==0.1.99